+++ apps/web/src/widgets/feed/model/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useFeedMain } from './useFeedMain';
 export { useFeedRecentSearchKeywordSection } from './useFeedRecentSearchKeywordSection';
+export { usePostListItem } from './usePostListItem';
